Struct wasi_cap_std_sync::TcpListener

pub struct TcpListener { /* private fields */ }

A TCP socket server, listening for connections.

This corresponds to std::net::TcpListener.

This TcpListener has no bind method. To bind it to a socket address, first obtain a Pool permitting the address, and then call Pool::bind_tcp_listener.

Implementations

impl TcpListener

pub fn from_std(std: TcpListener) -> TcpListener

Constructs a new instance of Self from the given std::net::TcpListener.

This grants access the resources the std::net::TcpListener instance already has access to.

pub fn local_addr(&self) -> Result<SocketAddr, Error>

Returns the local socket address of this listener.

This corresponds to std::net::TcpListener::local_addr.

pub fn try_clone(&self) -> Result<TcpListener, Error>

Creates a new independently owned handle to the underlying socket.

This corresponds to std::net::TcpListener::try_clone.

pub fn accept(&self) -> Result<(TcpStream, SocketAddr), Error>

Accept a new incoming connection from this listener.

This corresponds to std::net::TcpListener::accept.

pub fn incoming(&self) -> Incoming<'_>

Returns an iterator over the connections being received on this listener.

This corresponds to std::net::TcpListener::incoming.

pub fn set_ttl(&self, ttl: u32) -> Result<(), Error>

Sets the value for the IP_TTL option on this socket.

This corresponds to std::net::TcpListener::set_ttl.

pub fn ttl(&self) -> Result<u32, Error>

Gets the value of the IP_TTL option for this socket.

This corresponds to std::net::TcpListener::ttl.

pub fn take_error(&self) -> Result<Option<Error>, Error>

Gets the value of the SO_ERROR option on this socket.

This corresponds to std::net::TcpListener::take_error.

pub fn set_nonblocking(&self, nonblocking: bool) -> Result<(), Error>

Moves this TCP stream into or out of nonblocking mode.

This corresponds to std::net::TcpListener::set_nonblocking.